[Media Recorder] fix windows compile error for chrome 53 (+1 squashed commits)
Squashed commits:

[0af739d] [Media Recorder] add proprietary_codecs compile option,
add javascript resources,
adjust event dispatcher, enable audio capture for non screen capture streams,
pending data might contain negative value, cast it to int so we can get the actual audio delay,
[WIN] fix timestamp calculation, disable audio loopback on XP,
[WIN] force audio share on window content

+++ media/audio/audio_input_device.cc
@@ -344,7 +344,7 @@ void AudioInputDevice::AudioThreadCallback::Process(uint32_t pending_data) {
   // the audio delay measurement.
   capture_callback_->Capture(
       audio_bus,
-      buffer->params.hardware_delay_bytes / bytes_per_ms_,  // Delay in ms
+      static_cast<int>(buffer->params.hardware_delay_bytes) / bytes_per_ms_,  // Delay in ms
       buffer->params.volume, buffer->params.key_pressed);
 
   if (++current_segment_id_ >= total_segments_)
